moved functions to auxfunctions
I moved the album function to the auxfunctions file but I also implemented the image only channel function for #wholesome-memes-only and #art-share.

Not a big deal but I also changed every "ctx.channel.send" to "ctx.send" because that was a bit silly.
